requirements.txt
absl-py==0.9.0
alabaster==0.7.12
anaconda-client==1.2.2
asn1crypto==1.3.0
astor==0.8.0
attrs==19.3.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
clyent==1.2.1
cryptography==2.3.1
cycler==0.10.0
decorator==4.4.2
Django==2.0.2
gast==0.3.3
grpcio==1.14.1
h5py==2.8.0
idna==2.9
importlib-metadata==1.5.0
ipython-genutils==0.2.0
joblib==0.14.1
jsonschema==3.2.0
jupyter-core==4.6.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
mkl-fft==1.0.2
mkl-random==1.0.0.8
nbformat==5.0.4
numpy==1.16.2
scipy==1.2.1
matplotlib==3.1.1
olefile==0.46
opencv-python==4.2.0.32
Pillow==7.0.0
protobuf==3.11.3
psutil==5.6.0
pycparser==2.19
pyOpenSSL==19.0.0
pyparsing==2.4.6
pyrsistent==0.16.0
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3.1
QtPy==1.9.0
requests==2.23.0
ruamel-yaml==0.15.87
scikit-learn==0.22.1
six==1.14.0
sklearn==0.0
sqlparse==0.3.0
tensorboard==1.12.2
tensorflow==1.12.0
termcolor==1.1.0
torch==1.2.0
tornado==6.0.3
traitlets==4.3.3
urllib3==1.25.8
Werkzeug==1.0.0
zipp==2.2.0
